Added null reference check + vscodeworkspace
Added null check for gizmo drawing. This is done because the _rigidbody thing is only assigned at runtime, meaning that when you select the bee in the editor and the game isn't running it creates hundreds of null reference errors. This fixes that.
Plus a vscode workspace so i could name it in my sidebar