Dialogue can no longer overlap
Implimented event system so that is interacted changes only when the npc is finished talking to.
So what it does is on interaction, set isInteracted true.
This blocks interaction from the player, then it starts the function to talk with the npc.
Only after dialogu manager has verifed there is no more text to show, it sets isInteracted false, allowing interaction again.
This means that you can no longer interact with the same npc twice, causing an overlap in text